Ready for the trip to Irvine

+++ Applications/QuickITKApps/FillSTLMesh/FillSTLMesh.cxx
@@ -201,7 +201,7 @@ int main(int argc, char **argv)
     drawPolyData(pd);
 
   // Get the extents of the data
-  float *bounds = pd->GetBounds();  
+  double *bounds = pd->GetBounds();  
 
   cout << "STL mesh bounds: " << endl;
   cout << "   X : " << bounds[0] << " to " << bounds[1] << endl;
@@ -227,7 +227,7 @@ int main(int argc, char **argv)
     {
     for(unsigned int i=0;i<3;i++)
       {
-      float *x = pd->GetPoints()->GetPoint(pts[i]);
+      double *x = pd->GetPoints()->GetPoint(pts[i]);
       vtx[it] = (double *) malloc(3*sizeof(double));
       for(unsigned int j=0;j<3;j++)
         {
